+++ worker-wasm/core/src/minify.rs
@@ -1,5 +1,5 @@
 use std::error::Error;
-use crate::parser::JaqJsonFactory;
+use crate::jaq_json_factory::JaqJsonFactory;
 
 pub fn minify_json(input: &str) -> Result<String, Box<dyn Error>> {
     let val = crate::parser::parse_json(input.as_bytes(), JaqJsonFactory)?;
