+++ worker-wasm/core/src/node_json_factory.rs
@@ -1,0 +1,42 @@
+use jaq_json::Num;
+use crate::Node;
+use crate::parser::Factory;
+use crate::utils::determinate_variant;
+
+pub struct NodeJsonFactory;
+
+impl Factory<Node> for NodeJsonFactory {
+    fn null(&self) -> Node {
+        Node::Null
+    }
+
+    fn bool(&self, val: bool) -> Node {
+        Node::Boolean { value: val }
+    }
+
+    fn number(&self, n: Num) -> Node {
+        Node::Number {
+            value: n.to_string(),
+        }
+    }
+
+    fn string(&self, s: Vec<u8>) -> Node {
+        let string = String::from_utf8_lossy(&s);
+        Node::String {
+            value: string.to_string(),
+            variant: determinate_variant(string.trim()),
+        }
+    }
+
+    fn array(&self, arr: Vec<Node>) -> Node {
+        Node::array(arr)
+    }
+
+    fn object(&self, obj: Vec<(String, Node)>) -> Node {
+        Node::object(
+            obj.into_iter()
+                .map(|(k, v)| Node::property(&k, v))
+                .collect(),
+        )
+    }
+}
